+++ app/controllers/panorama_sampler_controller.rb
@@ -57,7 +57,7 @@ class PanoramaSamplerController < ApplicationController
     unless Panorama::Application.config.panorama_var_home_user_defined
       add_popup_message("There's no storage location defined for persistent config data!
 That means, this config data will be lost at next restart of Panorama backend application!
-To fix this, set environment variable PANORAMA_VAR_HOME to the desired location before starting the backend application.")
+To fix this, set config attribute PANORAMA_VAR_HOME to the desired location before starting the backend application.")
     end
 
     # warn if ASH is activated first time