+++ config/initializers/sql_logging.rb
@@ -1,2 +1,2 @@
 # Ensure logging of SQL statements if requested
-ActiveRecord::Base.logger = ActiveSupport::Logger.new(STDOUT) if ENV['PANORAMA_LOG_SQL'] && ENV['PANORAMA_LOG_SQL'].upcase == 'TRUE'
+ActiveRecord::Base.logger = ActiveSupport::Logger.new(STDOUT) if Panorama::Application.config.panorama_log_sql&.upcase == 'TRUE'